Resolve Failing Correct Data Test
Due to the fact that additional parameters were being merged into our request json, we were failing to validate our schema. In this commit we isolate the data in our request into a data key so that we can process only the data related JSON/hash and not ActionController related params. We also refactor our test to use as: jason which is a shortcut provided by ActionDispatch::IntegrationTest for setting JSON headers.